Various fixes and improvements
- Implement ToolProvider
- Change syntax for hexadecimal and binary integers. Add syntax for octal
  integers
- Support for negative indices in array "get" and "put" operators
- Do not expose NAME_SLASHED and NAME_QUOTED in grammar
- Change tests naming

+++ src/main/java/coneforest/psylla/core/PsyFormalArray.java
@@ -24,7 +24,7 @@ public interface PsyFormalArray<T extends PsyObject>
 	{
 		final PsyFormalArray<T> result=psyClone();
 		int length=result.length();
-		for(int i=0; i<(int)(length/2); i++)
+		for(int i=0; i<length/2; i++)
 		{
 			T o=result.get(i);
 			result.put(i, result.get(length-1-i));
